small bug fix in robot pool. docs, install debugging
